requirements.txt
Django
djangorestframework
python-decouple
PyJWT
psycopg2-binary
redis
celery
django-cors-headers
django-filter
djangorestframework-simplejwt
Pillow
factory-boy
channels
channels-redis
python-dotenv
gunicorn
whitenoise
